Implement shared memory mapping write-back support
- Updated the lseek function to accept 64-bit offsets in various syscall handlers to accommodate larger file sizes.
- Enhanced the sys_exit function to ensure shared memory mappings are written back to their files upon process exit.
- Introduced a new module for managing shared memory mappings, including functions for registering mappings, writing back dirty pages, and handling unmapping.
- Added support for new futex operations: FUTEX_REQUEUE, FUTEX_CMP_REQUEUE, and FUTEX_WAKE_OP, allowing for more complex synchronization scenarios.
- Implemented deferred wake handling to reduce I/O completion latency by allowing remote CPUs to drain deferred wakes immediately.
- Improved error handling and logging for operations involving shared memory mappings and file I/O.

+++ kernel/src/memory/swap.rs
@@ -202,7 +202,7 @@ pub fn free_slot(slot: SwapSlot) -> Result<(), SwapError> {
 pub fn write_page(slot: SwapSlot, page: &[u8; FRAME_SIZE]) -> Result<(), SwapError> {
     let (fd, off) = begin_slot_io(slot)?;
     let _write_guard = SWAP_WRITE_LOCK.lock();
-    let result = if crate::fs::vfs::lseek(fd, off as i32, 0).is_ok() {
+    let result = if crate::fs::vfs::lseek(fd, off as i64, 0).is_ok() {
         match crate::fs::vfs::write(fd, page) {
             Ok(written) if written == FRAME_SIZE => Ok(()),
             Ok(_) | Err(_) => Err(SwapError::Io),
